survey: check empty answers

+++ survey/views.py
@@ -98,27 +98,43 @@ class QuestionnaireView(View):
 
 		items = request.POST.items()
 		extra = {}
-		# generate the answer_dict for each question, and place in extra
+		# generate the answer_dict for each posted question, and place in extra
+		posted_ids = []
 		for item in items:
 			key, value = item[0], item[1]
 			if key.startswith('question_'):
-				answer = key.split("_", 2)
-			question = Question.objects.filter(sortid=answer[1], 
+				qssortid = key.split("_", 2)
+			question = Question.objects.filter(sortid=qssortid[1], 
+				questionset=questionset, 
+				questionset__questionnaire=questionnaire).first()
+			posted_ids.append(int(qssortid[1]))
+			if not question:
+				continue
+			extra[question] = value
+		#generate none for each empty quesiton, and place in extra
+		expected = questionset.questions()
+		empty_ids = []
+		for q in expected:
+			if q.sortid in posted_ids:
+				continue
+			empty_ids.append(q.sortid)
+		for q in empty_ids:
+			question = Question.objects.filter(sortid=q, 
 				questionset=questionset, 
 				questionset__questionnaire=questionnaire).first()
 			if not question:
 				continue
-			if (len(answer) == 2):
-				ans = value
-			else:
-				continue
-			extra[question] = ans
+			extra[question] = None
+
 		for question, ans in extra.items():
+			if ans in [None, '']:
+				kwargs = {'runid': runinfo.id}
+				return HttpResponseRedirect(reverse('survey:questionnaire', kwargs=kwargs))
 			answer = Answer()
 			answer.question = question
-			answer.answer = ans
 			answer.subject = runinfo.subject	
 			answer.runid = runinfo.pk	
+			answer.answer = ans
 			Answer.objects.filter(subject=runinfo.subject, 
 				runid=runinfo.pk, question=question).delete()
 			answer.save()
